+++ release_notes.md
@@ -4,6 +4,9 @@ FastLED3.1.1pre
 * Fix edge case bug w/HSV palette blending 
 * Fix power management issue w/parallel output
 * Use static_asserts for some more useful compile time errors around bad pins
+* Roll power management into FastLED.show/delay directly
+* Support for adafruit pixies on arduino type platforms that have SoftwareSerial
+  * TODO: support hardware serial on platforms that have it available
 
 FastLED3.1.0
 ============